refactor: clean up clippy warnings across Rust codebase
- Remove 17 redundant c_int-to-i32 type casts across FFI modules
- Collapse 10 nested if statements into combined conditions
- Replace 11 manual .max().min() patterns with .clamp()

+++ rust/neomacs-display/src/core/syntax_table.rs
@@ -739,7 +739,7 @@ pub fn scan_lists<T: TextSource + ?Sized>(
                         }
                         remaining -= 1;
                     }
-                    if !sexpflag && current_depth == depth - count as i32 + remaining as i32 {
+                    if !sexpflag && current_depth == depth - count + remaining {
                         // Done!
                     }
                 }
